Use rshared mount option for /var/tmp/mesos on Mesos slave
This is due to https://github.com/kubernetes/kubernetes/issues/31062

The fix is to pre-condition `/var/tmp/mesos` by (1) bind-mounting it
to itself and (2) `mount --make-rshared /var/tmp/mesos`, and on the
Mesos slave container pass the `rshared` option to the volume mount of
`/var/tmp/mesos`.  The Ansible modules `docker` and `docker_container`
do NOT support the `rshared` option on volume mounts, so we have to
use the `docker run` shell command.  This change loses much of the
sophistication of the `docker` module, regarding the variety of
initial states that will be correctly transformed into the desired
state.  We can either improve on this shell/ansible scripting and/or
improve the relevant module to support the `rshared` option.